requirements.txt
boto3==1.14.5
botocore==1.17.50
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.7
django-import-export==2.3.0
django-heroku==0.3.1
flake8==3.8.3
gunicorn==20.0.4
ipython==7.16.1
Pillow==7.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
whitenoise==5.2.0


